feat: 동시 로그인 방지 기능 추가
