+++ src/app/(main)/(protected)/myactivity/[id]/edit/page.tsx
@@ -1,6 +1,8 @@
+import { ipCheck } from '@/app/action/ip-check';
 import ActivityEditForm from '@/components/organisms/activity-edit-form/ActivityEditForm';
 import { getActivityDetails } from '@/queries/activities/get-activity-details';
 import { Metadata } from 'next';
+import { redirect } from 'next/navigation';
 
 export const metadata: Metadata = {
   title: '체험 수정 | Global Nomad',
@@ -11,6 +13,11 @@ interface Props {
   params: { id: number };
 }
 export default async function ActivityEditPage({ params }: Props) {
+  const isEqualIpUser = await ipCheck();
+  if (isEqualIpUser) {
+    redirect('/calendar');
+  }
+
   const activityData = await getActivityDetails(params.id);
   return <ActivityEditForm stateData={activityData} />;
 }
